% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compData.R
\name{compData}
\alias{compData}
\title{Creation of Comparison Data}
\usage{
compData(df1, df2, types = NULL, breaks = c(0.001, 0.25, 0.5))
}
\arguments{
\item{df1, df2}{two data frames to be linked, containing the same number of columns,
with the same variable names and classes, in the same order.  The columns represent 
the comparison fields to be used for the linkage.  Currently, only columns of class
\code{character} and \code{factor} are supported.  Without loss of generality, 
    \code{df1} is assumed to have no less records than \code{df2}.}

\item{types}{a vector of characters indicating the comparison types per comparison field.  The options
are: \code{"Lev"} for comparisons based on the Levenshtein distance normalized to \eqn{[0,1]}, with \eqn{0}  
indicating no disagreement and \eqn{1} indicating maximum disagreement; and 
\code{"bin"} for binary comparisons (agreement/disagreement). 
    The default is \code{"Lev"} for columns of class \code{character} and \code{"bin"} for columns of class \code{factor}.}

\item{breaks}{for comparisons based on the normalized Levenshtein distance, a vector of length \eqn{L} of break 
points for the interval \eqn{[0,1]} to obtain \eqn{L+1} levels of disagreement.
The default is \code{breaks=c(.001,.25,.5)}.}
}
\value{
a list containing: 
\describe{
  \item{\code{comparisons}}{
						matrix with \code{n1*n2} rows, where the comparison pattern for record pair \eqn{(i,j)}
					appears in row \code{(j-1)*n1+i}, for \eqn{i} in \eqn{{1,\dots,n1}}, and \eqn{j} in \eqn{{1,\dots,n2}}.  A comparison field with \eqn{L+1} levels of disagreement, 
					is represented by \eqn{L+1} columns of TRUE/FALSE indicators.  Missing comparisons are coded as FALSE,
					which is justified under an assumption of ignorability of the missing comparisons, see Sadinle (2017).
					}
  \item{\code{n1,n2}}{the datafile sizes, \code{n1 = nrow(df1)} and \code{n2 = nrow(df2)}.}
  \item{\code{nDisagLevs}}{a vector containing the number of levels of
				 	disagreement per comparison field.}
}
}
\description{
Create comparison vectors for all pairs of records coming from 
two datafiles to be linked.
}
\examples{
data(twoFiles)

myCompData <- compData(df1, df2, types=c("Lev","Lev","bin","bin"), breaks=c(.001,.25,.5))

## same as 
myCompData <- compData(df1, df2)
}
\references{
Mauricio Sadinle (2017). Bayesian Estimation of Bipartite Matchings for Record Linkage. \emph{Journal of the
American Statistical Association} 112(518), 600-612.
}
